open diplomatic dialog on click on leader
