+++ SmartColonyMacOSUILibrary/SmartColonyMacOSUILibrary/Views/Custom/LeaderView.swift
@@ -6,50 +6,6 @@
 //
 
 import SwiftUI
-import SmartAILibrary
-import SmartAssets
-
-protocol LeaderViewModelDelegate: AnyObject {
-
-    func clicked(on leaderType: LeaderType)
-}
-
-class LeaderViewModel: ObservableObject {
-
-    let id: UUID = UUID()
-
-    let leaderType: LeaderType
-
-    weak var delegate: LeaderViewModelDelegate?
-
-    init(leaderType: LeaderType) {
-
-        self.leaderType = leaderType
-    }
-
-    func image() -> NSImage {
-
-        return ImageCache.shared.image(for: self.leaderType.iconTexture())
-    }
-
-    func clicked() {
-
-        self.delegate?.clicked(on: self.leaderType)
-    }
-}
-
-extension LeaderViewModel: Hashable {
-
-    static func == (lhs: LeaderViewModel, rhs: LeaderViewModel) -> Bool {
-
-        return lhs.id == rhs.id
-    }
-
-    func hash(into hasher: inout Hasher) {
-
-        hasher.combine(self.id)
-    }
-}
 
 struct LeaderView: View {
 
@@ -62,7 +18,7 @@ struct LeaderView: View {
 
             Image(nsImage: self.viewModel.image())
                 .resizable()
-                .frame(width: 32, height: 32, alignment: .topLeading)
+                .frame(width: 42, height: 42, alignment: .topLeading)
                 .onTapGesture {
                     self.viewModel.clicked()
                 }
